requirements.txt
fastapi
psycopg2
SQLAlchemy
uvicorn
python-jose
passlib